requirements.txt
APScheduler==3.7.0	
aiohttp==3.7.4
beautifulsoup4==4.9.3
discord.py==2.4.0
humanfriendly==9.1
matplotlib==3.9.0
psycopg2==2.9.5
python-dotenv==0.15.0
psutil==5.8.0